Convert parseLocals() functions to use the return value for the result.
It is modern practice to prefer return values over pass-by-reference, since
compilers can optimize the code sufficiently well.

While we are here, avoid a copy in KDebugger::parseLocals(). The names of
duplicated variables must be adjusted, but this does not require to copy
the values around: just modify the list in place.