Hold breakpoint objects in std::unique_ptr<>.
Modern C++ should not use plain pointers. Use std::unique_ptr<> instead to
indicated ownership.

Pass a reference instead of a pointer to executeBreakpoint(), because the
pointer is non-owning and the use of the pointed-to value is only like a
value, not a pointer. Additionally, convert a use of iterators at the call
site in handleRunCommands() to a range-based for-loop. This makes the code
more idiomatic.